requirements.txt
streamlit>=1.31.0
open-clip-torch>=2.24.0
pybioclip>=2.1.1
torch>=2.1.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
dashscope>=1.20.0
faiss-cpu>=1.7.4
numpy>=1.24.0
filelock>=3.13.0
requests>=2.31.0
ultralytics>=8.2.0
pytest>=8.0.0
